:sparkles: Allow to specify a different name for field in serialized/deserialized json for a POJO field
If property 'msgField' is set on an enpoint parameter field, then expected name for corresponding JSON is this value and not the 'name' property.
This is useful to provide a readable meaningful 'name' to POJO fields, when corresponding JSON field  is not, for instance a single letter. This is often the case in websocket stream messages to reduce bandwidth