💥JavaxNetRestEnpoint#getBody() now returns JSON encoded value of its request data
This body will not be written if request method is GET as such request must not carry a body.
This triggers a change in BinancePrivateApiRestEnpoint that now override this method to return null, since this is how this API works, providing parameters as URL params even in POST requests. This change has been done to for public REST APIs though it may not be needed since public API endpoints are only GET... but this is more consistent.