💥 Changed generated POJOs (for both request and incoming response messages) to use expanded type instead of primitive types
This means Integer, Long, Boolean fields generated instead of int, long, boolean. The reason for this is to distinguish not set fields versus fields set with a default value. Now, any null value request parameter field will not be serialized in URL parameter, this was not possible for optional 'timestamp' parameters for instance.