:sparkles: Adds placeholder resolution utilities
Introduces a new `PlaceHolderResolver` interface and a default implementation
using `StringSubstitutor`. This allows for easy placeholder replacement in
strings using a map of key-value pairs.

Also, adds a utility method to `CollectionUtil` to create maps from
alternating key and value pairs. This simplifies the creation of maps
required for the placeholder resolution.

Refactors `EncodingUtil` to leverage `CollectionUtil.createMap` and
introduces a separate method for argument substitution using a map.