:sparkles: Adds all conditional JSON field writing in JsonUtil
Adds methods to write boolean, long, BigDecimal, and integer fields to a JSON generator only if the values are not null or, in the case of boolean, false.

This avoids writing null or default values to the JSON output, which can be useful for creating cleaner and more compact JSON representations.