:recycle: :sparkles: Refactored 'constants' declaration and generation
'Constant' descriptor object is now recursive and may contain a list of nested constants in which case it is a group of nested constants.

No more constants at API group level this was not so  useful  for  wrapper users to guess where a constant is defined. Constants are all defined at 'exchange' level but may be grouped through 'grouping' feature