:sparkles: Introduces 'requestHasBody' REST endpoint descriptor property
This allows to specify wether the endpoint request should be serialized as (JSON) body or not.
Should be usually left unset,  so this behavior is derived from HTTP method. Can be useful in case  specific API choose to provide for instance DELETE request as body instead of URL query/path parameters.